refactor and adding CreateProject func
