+++ pkg/client/waypoint_server_test.go
@@ -1,0 +1,1 @@
+package client
